Move application bundle client to pkg/internal/
To encourage a uniform way of accessing application bundles, I'm
moving this package out of pkg/server. This makes it clear that can be
used from both the server handles (which create K8s objects) and the
provisioners (which react to those objects).

This will give me a single place to insist on providing a namespace
for application bundle lookups, whether it's list or individual
objects.