Supply an applicationbundle.Client explicitly
Many operations need an *applicationbundle.Client to list or get
application bundles. Since I want them to be constructed with a
specific namespace, I have

 - required the namespace in the constructor; and,
 - centralised the construction in the outer server handler, and
   passed the client through explicitly to sub-handlers when they
   need it.

This reduces the effort needed to change how application bundles are
looked up, as well as the opportunities for getting it wrong
somewhere.

The behaviour is still to use the application bundles from the
controller namespace -- this is given to both the provisioners (in the
context) and to the server handler (as a field value).