Change meaning of 'prefix' to be "append with a slash"
The flag `--node-selector-label-is-prefix` changes how the value of
`--node-selector-label` is used, and it's easy to end up with e.g.,
`foobarvcluster123` instead of what you probably wanted,
`foobar/vcluster123`.

So, instead of being infinitely flexible and literal, just add the
`"/"` for people (and say that's what will happen in the help text).