New flag for VirtualKubernetesCluster controller: --node-selector-label (#249)
When creating VirtualKubernetesClusters especially, it will sometimes
be useful to limit the nodes used by the vCluster, with a node
selector. The setting `fromHost.nodes.selector.labels` both limits
which nodes appear from inside the vcluster, and which nodes pods can
be scheduled on.

I've represented this as a flag `--node-selector-label` -- label
because the _value_ is going to be particular to the vcluster. It has
a counterpart in the Helm chart values.yaml.

Using a node selector of say, `vcluster/abc123: ""` is more flexible
than using `vcluster: abc123`, because it allows a node to be included
in more than one vcluster, without making anything else difficult
(other than enforcing single tenancy -- but that can be done
elsewhere).

The flag `--node-selector-label-is-prefix` changes how the value of
`--node-selector-label` is used, and it's easy to end up with e.g.,
`foobarvcluster123` instead of what you probably wanted,
`foobar/vcluster123`.

So, instead of being infinitely flexible and literal, just add the
`"/"` for people (and say that's what will happen in the help text).