Include workload pool specs in chart values (#250)
It's useful for any virtual cluster chart to be able to create objects
based on the workload pool specs, which aren't otherwise processed by
the controller.

This only really makes sense in the context of #248 (which made the
vcluster chart a subchart, leaving room for the _parent_ chart to have
its own templates).