Scope virtual cluster deploy secrets to org
When creating a virtual cluster, the provisioner uses ProvisionOn to
install the vcluster chart on the region (underlay) cluster. The
remote provisioner ultimately creates a Secret containing the
kubeconfig so that ArgoCD can install to the remote cluster; that
secret has a name _field_ (which ArgoCD uses as its handle), and a
metadata.name (which is the name of the Secret so far as Kubernetes is
concerned.

Previously, these did not line up: the name field was generated from
the VirtualKubernetesCluster labels, while the metadata.name was the
host with an optional prefix. Since the host was the same each time,
each new VirtualKubernetesCluster would overwrite the Secret with its
own name field. To ArgoCD this meant the deployment target of each
virtual cluster would disappear, and it couldn't sync that virtual
cluster.

The solution here is to line the name field and the Secret name up, so
that:

 - if it's the same target, the name stays the same;
 - if it's a different target cluster, it gets its own Secret.

I've scoped these to the organisation, conservatively -- if a region
is used from more than one organisation, they will all get their own
secret.