Use the same remote namespace in provisioner and server (#269)
My previous change in #261 gave a prefix to the namespace that a
virtual cluster is deployed to. But the server also uses that
namespace, to retrieve the kubeconfig secret for the vcluster.

These relied on coincidence, which I broke. Instead, use a common func
to calculate the namespace. This still isn't perfect -- ideally I
suppose it would find the ArgoCD application and get the namepace from
there. But this change removes the coincidence, at least.