Fix Remote Cluster Aliasing (#277)
When you provision on a remote cluster the expectation is that the
provisioning resource owns that remote cluster configuration, and it can
delete it when finished with it.  What we ended up with here is a single
secret defined for all virtual clusters on the same remote, thus
deleting a virtual cluster removed the remote secret for all other
virtual clusters.  What was more insidious was multiple deletions at the
same time would lead to resource leaks as the remote code goes "oh you
don't exist, I must have cleaned you up" and for idempotency ignores the
children and carries on, leaving the children alive and well.